to revise water entry case
