+++ src/shared/sphinxsys_system/sph_system.cpp
@@ -54,9 +54,9 @@ void SPHSystem::handleCommandlineOptions(int ac, char *av[])
 
         po::options_description desc("Allowed options");
         desc.add_options()("help", "produce help message");
-        desc.add_options()("r", po::value<bool>(), "Particle relaxation.");
-        desc.add_options()("i", po::value<bool>(), "Particle reload from input file.");
-        desc.add_options()("rt", po::value<bool>(), "Regression test.");
+        desc.add_options()("relax", po::value<bool>(), "Particle relaxation.");
+        desc.add_options()("reload", po::value<bool>(), "Particle reload from input file.");
+        desc.add_options()("regression", po::value<bool>(), "Regression test.");
         desc.add_options()("state_recording", po::value<bool>(), "State recording in output folder.");
         desc.add_options()("restart_step", po::value<int>(), "Run form a restart file.");
 
@@ -70,11 +70,11 @@ void SPHSystem::handleCommandlineOptions(int ac, char *av[])
             exit(0);
         }
 
-        if (vm.count("r"))
+        if (vm.count("relax"))
         {
-            run_particle_relaxation_ = vm["r"].as<bool>();
+            run_particle_relaxation_ = vm["relax"].as<bool>();
             std::cout << "Particle relaxation was set to "
-                      << vm["r"].as<bool>() << ".\n";
+                      << vm["relax"].as<bool>() << ".\n";
         }
         else
         {
@@ -82,11 +82,11 @@ void SPHSystem::handleCommandlineOptions(int ac, char *av[])
                       << run_particle_relaxation_ << ").\n";
         }
 
-        if (vm.count("i"))
+        if (vm.count("reload"))
         {
-            reload_particles_ = vm["i"].as<bool>();
+            reload_particles_ = vm["reload"].as<bool>();
             std::cout << "Particle reload from input file was set to "
-                      << vm["i"].as<bool>() << ".\n";
+                      << vm["reload"].as<bool>() << ".\n";
         }
         else
         {
@@ -94,11 +94,11 @@ void SPHSystem::handleCommandlineOptions(int ac, char *av[])
                       << reload_particles_ << ").\n";
         }
 
-        if (vm.count("rt"))
+        if (vm.count("regression"))
         {
-            generate_regression_data_ = vm["rt"].as<bool>();
+            generate_regression_data_ = vm["regression"].as<bool>();
             std::cout << "Generate regression test data set was set to "
-                      << vm["rt"].as<bool>() << ".\n";
+                      << vm["regression"].as<bool>() << ".\n";
         }
         else
         {
