+++ tests/user_examples/test_2d_stretching/CMakeLists.txt
@@ -22,14 +22,7 @@ target_link_libraries(${PROJECT_NAME} extra_sources_2d)
 set_target_properties(${PROJECT_NAME} PROPERTIES VS_DEBUGGER_WORKING_DIRECTORY "${EXECUTABLE_OUTPUT_PATH}")
 
  
-if(${CMAKE_SYSTEM_NAME} MATCHES "Windows")
-    add_test(NAME ${PROJECT_NAME}_particle_relaxation COMMAND ${PROJECT_NAME} --r=true
+add_test(NAME ${PROJECT_NAME}_particle_relaxation COMMAND ${PROJECT_NAME} --relax=true --state_recording=${TEST_STATE_RECORDING}
                      WORKING_DIRECTORY ${EXECUTABLE_OUTPUT_PATH})
-    add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME} --r=false --i=true
+add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME}  --reload=true --state_recording=${TEST_STATE_RECORDING}
                      WORKING_DIRECTORY ${EXECUTABLE_OUTPUT_PATH})
-else()
-    file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/run_test.sh
-            DESTINATION ${EXECUTABLE_OUTPUT_PATH})
-    add_test(NAME ${PROJECT_NAME} COMMAND bash ${EXECUTABLE_OUTPUT_PATH}/run_test.sh
-            WORKING_DIRECTORY ${EXECUTABLE_OUTPUT_PATH})
-endif()
