--- tests/user_examples/xj_stretching/run_test.sh
@@ -1,2 +1,0 @@
-./xj_stretching --r=true
-./xj_stretching --r=false --i=true
